Convert Result to a template class.
The Result class is now a template class so it can return 'data' that matches
the value type of the ValuStor class. Previously this was always a std::string.
Added a size to the Result class so that byte arrays and char* can be properly sized.

Fixed a bug where the backlog was not using the templated types for ValuStor.

Added some basic sanity tests to the example program.