Initialize the DB connection in the backlog thread.
The database connection initialization is now done by the backlog thread.
This fixes the issue where values could get stuck in the backlog.
This change allows us to remove a mutex from the code and replace it
with polling an atomic instead.

Fixed a pointer initialization race condition.
On x64 platforms these writes are inherently atomic, but this would
have broken portability.
Calls to store() and retrieve() no longer block on an uninitialized state.
Instead, the calls fail immediately. In the case of store() where the
backlog is used, the requests are immediately added to the backlog queue.
This should improve the performance of these functions and prevent
them from blocking before the initialization completes.